move tags tests
